Move posix_util to utils/ dir.
Also removes "from gslib import util" that I forgot to remove from the
previous commit.
