+++ gslib/commands/rsync.py
@@ -52,26 +52,6 @@ from gslib.hashing_helper import SLOW_CRCMOD_RSYNC_WARNING
 from gslib.hashing_helper import SLOW_CRCMOD_WARNING
 from gslib.metrics import LogPerformanceSummaryParams
 from gslib.plurality_checkable_iterator import PluralityCheckableIterator
-from gslib.posix_util import ATIME_ATTR
-from gslib.posix_util import ConvertDatetimeToPOSIX
-from gslib.posix_util import ConvertModeToBase8
-from gslib.posix_util import DeserializeFileAttributesFromObjectMetadata
-from gslib.posix_util import GID_ATTR
-from gslib.posix_util import InitializeUserGroups
-from gslib.posix_util import MODE_ATTR
-from gslib.posix_util import MTIME_ATTR
-from gslib.posix_util import NA_ID
-from gslib.posix_util import NA_MODE
-from gslib.posix_util import NA_TIME
-from gslib.posix_util import NeedsPOSIXAttributeUpdate
-from gslib.posix_util import ParseAndSetPOSIXAttributes
-from gslib.posix_util import POSIXAttributes
-from gslib.posix_util import SerializeFileAttributesToObjectMetadata
-from gslib.posix_util import UID_ATTR
-from gslib.posix_util import ValidateFilePermissionAccess
-from gslib.posix_util import WarnFutureTimestamp
-from gslib.posix_util import WarnInvalidValue
-from gslib.posix_util import WarnNegativeAttribute
 from gslib.seek_ahead_thread import SeekAheadResult
 from gslib.sig_handling import GetCaughtSignals
 from gslib.sig_handling import RegisterSignalHandler
@@ -86,6 +66,26 @@ from gslib.utils.constants import UTF8
 from gslib.utils.metadata_util import CreateCustomMetadata
 from gslib.utils.metadata_util import GetValueFromObjectCustomMetadata
 from gslib.utils.metadata_util import ObjectIsGzipEncoded
+from gslib.utils.posix_util import ATIME_ATTR
+from gslib.utils.posix_util import ConvertDatetimeToPOSIX
+from gslib.utils.posix_util import ConvertModeToBase8
+from gslib.utils.posix_util import DeserializeFileAttributesFromObjectMetadata
+from gslib.utils.posix_util import GID_ATTR
+from gslib.utils.posix_util import InitializeUserGroups
+from gslib.utils.posix_util import MODE_ATTR
+from gslib.utils.posix_util import MTIME_ATTR
+from gslib.utils.posix_util import NA_ID
+from gslib.utils.posix_util import NA_MODE
+from gslib.utils.posix_util import NA_TIME
+from gslib.utils.posix_util import NeedsPOSIXAttributeUpdate
+from gslib.utils.posix_util import ParseAndSetPOSIXAttributes
+from gslib.utils.posix_util import POSIXAttributes
+from gslib.utils.posix_util import SerializeFileAttributesToObjectMetadata
+from gslib.utils.posix_util import UID_ATTR
+from gslib.utils.posix_util import ValidateFilePermissionAccess
+from gslib.utils.posix_util import WarnFutureTimestamp
+from gslib.utils.posix_util import WarnInvalidValue
+from gslib.utils.posix_util import WarnNegativeAttribute
 from gslib.utils.rsync_util import DiffAction
 from gslib.utils.rsync_util import RsyncDiffToApply
 from gslib.utils.system_util import IS_WINDOWS
