+++ gslib/utils/posix_util.py
